KAFKA-3618: Address review comments
