+++ core/src/test/scala/unit/kafka/server/ApiVersionsRequestTest.scala
@@ -18,8 +18,25 @@
 package kafka.server
 
 import org.apache.kafka.common.protocol.ApiKeys
+import org.apache.kafka.common.requests.ApiVersionsResponse.ApiVersion
 import org.apache.kafka.common.requests.{ApiVersionsRequest, ApiVersionsResponse}
+import org.junit.Assert._
 import org.junit.Test
+
+import scala.collection.JavaConversions._
+
+object ApiVersionsRequestTest {
+  def validateApiVersionsResponse(apiVersionsResponse: ApiVersionsResponse) {
+    assertEquals("API keys in ApiVersionsResponse must match API keys supported by broker.", ApiKeys.values.length, apiVersionsResponse.apiVersions.size)
+    for (expectedApiVersion: ApiVersion <- ApiVersionsResponse.apiVersionsResponse.apiVersions) {
+      val actualApiVersion = apiVersionsResponse.apiVersion(expectedApiVersion.apiKey)
+      assertNotNull(s"API key ${actualApiVersion.apiKey} is supported by broker, but not received in ApiVersionsResponse.", actualApiVersion)
+      assertEquals("API key must be supported by the broker.", expectedApiVersion.apiKey, actualApiVersion.apiKey)
+      assertEquals(s"Received unexpected min version for API key ${actualApiVersion.apiKey}.", expectedApiVersion.minVersion, actualApiVersion.minVersion)
+      assertEquals(s"Received unexpected max version for API key ${actualApiVersion.apiKey}.", expectedApiVersion.maxVersion, actualApiVersion.maxVersion)
+    }
+  }
+}
 
 class ApiVersionsRequestTest extends BaseRequestTest {
 
@@ -28,7 +45,7 @@ class ApiVersionsRequestTest extends BaseRequestTest {
   @Test
   def testApiVersionsRequest() {
     val apiVersionsResponse = sendApiVersionsRequest(new ApiVersionsRequest, 0)
-    validateApiVersionsResponse(apiVersionsResponse)
+    ApiVersionsRequestTest.validateApiVersionsResponse(apiVersionsResponse)
   }
 
   private def sendApiVersionsRequest(request: ApiVersionsRequest, version: Short): ApiVersionsResponse = {
@@ -36,4 +53,3 @@ class ApiVersionsRequestTest extends BaseRequestTest {
     ApiVersionsResponse.parse(response)
   }
 }
-
